+++ service/src/test/resources/transparencytestdata.json
@@ -1,0 +1,107 @@
+[
+    {
+        "name": "Starwit Anomaly Classifier",
+        "description": "Classifies Anomalies",
+        "version": "0.0.1",
+        "applicationIdentifier": "traffic",
+        "useAI": true,
+        "model": {
+            "name": "Starwit Classifier finetuned model",
+            "version": "1.0.0",
+            "lastDeployment": "2025-01-25T12:36:50.277768993+01:00",
+            "modelType": "LLM",
+            "modelLink": "https://example.com/",
+            "publicTrainingData": true,
+            "linkToPublicTrainingData": "http://starwit.de"
+        },
+        "actionTypes": [],
+        "decisionTypes": [],
+        "sBOMLocation": {
+            "backend": {
+                "url": "http://localhost:8080/sbom-samples/aic-sbom-backend.json"
+            }
+        }
+    },
+    {
+        "name": "Starwit GeoMerger",
+        "description": "Combines multiple detections of the same object into one geo representation",
+        "version": "0.0.1",
+        "applicationIdentifier": "traffic",
+        "useAI": false,
+        "model": {
+            "name": "Starwit Classifier finetuned model",
+            "version": "1.0.0",
+            "lastDeployment": "2025-01-25T12:36:50.277768993+01:00",
+            "modelType": "LLM",
+            "modelLink": "https://example.com/",
+            "publicTrainingData": true,
+            "linkToPublicTrainingData": "http://starwit.de"
+        },
+        "actionTypes": [],
+        "decisionTypes": [],
+        "sBOMLocation": {
+            "backend": {
+                "url": "http://localhost:8080/sbom-samples/aic-sbom-backend.json"
+            }
+        }
+    },
+    {
+        "name": "Anomaly Detection2",
+        "description": "Detects traffic anomalies",
+        "version": "0.0.1",
+        "applicationIdentifier": "traffic",
+        "useAI": true,
+        "model": {
+            "name": "Kernel density estimate",
+            "version": "1.0.0",
+            "lastDeployment": "2025-01-25T12:36:50.277768993+01:00",
+            "modelType": "MISC",
+            "modelLink": "https://example.com/",
+            "publicTrainingData": false,
+            "linkToPublicTrainingData": "https://example.com/"
+        },
+        "actionTypes": [],
+        "decisionTypes": [],
+        "sBOMLocation": {
+            "backend": {
+                "url": "http://localhost:8080/sbom-samples/aic-sbom-backend.json"
+            }
+        },
+        "successors": [
+            {
+                "name": "Starwit Anomaly Classifier"
+            }
+        ]
+    },
+    {
+        "name": "Starwit Object Tracker",
+        "description": "Tracks moving objects",
+        "version": "0.0.1",
+        "applicationIdentifier": "traffic",
+        "useAI": true,
+        "model": {
+            "name": "Yolov8",
+            "version": "8.0.0",
+            "lastDeployment": "2025-01-25T12:36:50.277768993+01:00",
+            "modelType": "COMPUTERVISION",
+            "modelLink": "https://example.com/",
+            "publicTrainingData": true,
+            "linkToPublicTrainingData": "http://starwit.de"
+        },
+        "actionTypes": [],
+        "decisionTypes": [],
+        "sBOMLocation": {
+            "backend": {
+                "url": "http://localhost:8080/sbom-samples/aic-sbom-backend.json"
+            }
+        },
+        "successors": [
+            {
+                "name": "Anomaly Detection2"
+            },
+            {
+                "name": "Starwit GeoMerger"
+            }
+        ]
+    }
+]
